minor changes to example build file Inspection run, minor things, exceptions not thrown, final and such DatabaseConnectionImpl.commit() and rollback() now throw exceptions as per method contract instead of ignoring them TextFieldPlus removed SizedDocument added, serves the purpose of the class removed above IntField and DoubleField refactored to accomodate the above RegexFilterCriteria refactored a bit
git-svn-id: svn://svn.hafro.is/jminor/trunk@6911 51816742-9f83-410d-85ed-e7bcf7042953

+++ src/org/jminor/common/db/Databases.java
@@ -257,12 +257,12 @@ public final class Databases {
       final long current = System.currentTimeMillis();
       final double seconds = (current - queriesPerSecondTime) / 1000d;
       if (seconds > 0) {
-        queriesPerSecond = (int) (queriesPerSecondCounter / (double) seconds);
-        selectsPerSecond = (int) (selectsPerSecondCounter / (double) seconds);
-        insertsPerSecond = (int) (insertsPerSecondCounter / (double) seconds);
-        deletesPerSecond = (int) (deletesPerSecondCounter / (double) seconds);
-        updatesPerSecond = (int) (updatesPerSecondCounter / (double) seconds);
-        undefinedPerSecond = (int) (undefinedPerSecondCounter / (double) seconds);
+        queriesPerSecond = (int) (queriesPerSecondCounter / seconds);
+        selectsPerSecond = (int) (selectsPerSecondCounter / seconds);
+        insertsPerSecond = (int) (insertsPerSecondCounter / seconds);
+        deletesPerSecond = (int) (deletesPerSecondCounter / seconds);
+        updatesPerSecond = (int) (updatesPerSecondCounter / seconds);
+        undefinedPerSecond = (int) (undefinedPerSecondCounter / seconds);
         queriesPerSecondCounter = 0;
         selectsPerSecondCounter = 0;
         insertsPerSecondCounter = 0;
